Fix Windows MSVC CI
Use portable C17/C++20 standards and fix static MSVC dependency propagation.

+++ include/siecs/cpp/query.hpp
@@ -201,7 +201,7 @@ inline void append_term(
     assert(term_index < ECS_QUERY_TERM_CAPACITY);
     desc.terms[term_index++] = {
         .id = id,
-        .access = access,
+        .access = static_cast<uint32_t>(access),
     };
 }
